Ispravlenije predupreždenij sborki.

+++ src/renderer/wvb.cpp
@@ -278,8 +278,8 @@ r_WorldVBOCluster& r_WVB::GetCluster( int longitude, int latitude )
 	int x= (longitude - cpu_cluster_matrix_coord_[0]) / cluster_size_[0];
 	int y= (latitude  - cpu_cluster_matrix_coord_[1]) / cluster_size_[1];
 
-	H_ASSERT( x >= 0 && x < cluster_matrix_size_[0] );
-	H_ASSERT( y >= 0 && y < cluster_matrix_size_[1] );
+	H_ASSERT( x >= 0 && x < (int)cluster_matrix_size_[0] );
+	H_ASSERT( y >= 0 && y < (int)cluster_matrix_size_[1] );
 
 	return *( cpu_cluster_matrix_[ x + y * cluster_matrix_size_[0] ] );
 }
@@ -292,8 +292,8 @@ r_WorldVBOClusterSegment& r_WVB::GetClusterSegment( int longitude, int latitude 
 	int cluster_x= d_lon / cluster_size_[0];
 	int cluster_y= d_lat / cluster_size_[1];
 
-	H_ASSERT( cluster_x >= 0 && cluster_x < cluster_matrix_size_[0] );
-	H_ASSERT( cluster_y >= 0 && cluster_y < cluster_matrix_size_[1] );
+	H_ASSERT( cluster_x >= 0 && cluster_x < (int)cluster_matrix_size_[0] );
+	H_ASSERT( cluster_y >= 0 && cluster_y < (int)cluster_matrix_size_[1] );
 
 	int segment_x= d_lon - cluster_x * cluster_size_[0];
 	int segment_y= d_lat - cluster_y * cluster_size_[1];
